Create a SolidusAfterpay::BaseController to handle common functionalities
