+++ app/controllers/solidus_afterpay/callbacks_controller.rb
@@ -1,12 +1,7 @@
 # frozen_string_literal: true
 
 module SolidusAfterpay
-  class CallbacksController < ::Spree::BaseController
-    protect_from_forgery except: [:confirm, :cancel]
-
-    rescue_from ::ActiveRecord::RecordNotFound, with: :resource_not_found
-    rescue_from ::CanCan::AccessDenied, with: :unauthorized
-
+  class CallbacksController < SolidusAfterpay::BaseController
     before_action :ensure_afterpay_order_token_presence, only: :confirm
     before_action :ensure_order_not_completed, only: :confirm
 
@@ -51,24 +46,6 @@ module SolidusAfterpay
 
     def afterpay_order_token
       params[:orderToken] || params[:order_token]
-    end
-
-    def order_token
-      cookies.signed[:guest_token]
-    end
-
-    def resource_not_found
-      respond_to do |format|
-        format.html { redirect_to spree.cart_path, notice: I18n.t('solidus_afterpay.resource_not_found') }
-        format.json { render json: { error: I18n.t('solidus_afterpay.resource_not_found') }, status: :not_found }
-      end
-    end
-
-    def unauthorized
-      respond_to do |format|
-        format.html { redirect_to spree.cart_path, notice: I18n.t('solidus_afterpay.unauthorized') }
-        format.json { render json: { error: I18n.t('solidus_afterpay.unauthorized') }, status: :unauthorized }
-      end
     end
 
     def ensure_afterpay_order_token_presence
