+++ app/controllers/solidus_afterpay/checkouts_controller.rb
@@ -1,12 +1,7 @@
 # frozen_string_literal: true
 
 module SolidusAfterpay
-  class CheckoutsController < ::Spree::BaseController
-    protect_from_forgery except: [:create]
-
-    rescue_from ::ActiveRecord::RecordNotFound, with: :resource_not_found
-    rescue_from ::CanCan::AccessDenied, with: :unauthorized
-
+  class CheckoutsController < SolidusAfterpay::BaseController
     def create
       authorize! :update, order, order_token
 
@@ -47,18 +42,6 @@ module SolidusAfterpay
       params[:redirect_cancel_url] || solidus_afterpay.callbacks_cancel_url(
         order_number: order.number, payment_method_id: payment_method.id
       )
-    end
-
-    def order_token
-      cookies.signed[:guest_token]
-    end
-
-    def resource_not_found
-      render json: { error: I18n.t('solidus_afterpay.resource_not_found') }, status: :not_found
-    end
-
-    def unauthorized
-      render json: { error: I18n.t('solidus_afterpay.unauthorized') }, status: :unauthorized
     end
   end
 end
